meta-lhg-integration: add doxygen recipe from meta-arago-extras
The Ducati decoding/encoding gstreamer plugin needs IPUMM firmware to
be loaded into the IPU units, or it will fail to work otherwise.

To build this firmware one needs doxygen. Here is the dependencies chain:
ipumm-fw -> ti-ipc-rtos -> doxygen-native

doxygen is not present in the generic layers we use for RPB and LHG builds,
so copy it from meta-arago-extras.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>